+++ NEWS.md
@@ -1,6 +1,8 @@
 MathOptInterface (MOI) release notes
 ====================================
 
+- Added `GeoMeantoRelEntrBridge` to bridge a geomean constraint to a relative entropy constraint.
+
 v0.9.17 (September 21, 2020)
 ---------------------------
 
@@ -112,7 +114,6 @@ v0.9.11 (February 21, 2020)
 v0.9.10 (January 31, 2020)
 ---------------------
 
-- Added `GeoMeantoRelEntrBridge` to bridge a geomean constraint to a relative entropy constraint.
 - Added `OptimizerWithAttributes` grouping an optimizer constructor and a list
   of optimizer attributes (#1008).
 - Added `RelativeEntropyCone` with corresponding bridge into exponential cone